requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
google-generativeai>=0.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
numpy>=1.24.0
